Refactor: Coffee 타입 분리
- CoffeeShop에서 Coffee타입을 분리
